requirements.txt
grpcio==1.71.0
grpcio-tools==1.71.0
opencv-python==4.11.0.86
protobuf==5.29.4
pytesseract==0.3.13
torch==2.7.0
ultralytics==8.3.144
